+++ dart/geobase/lib/coordinates.dart
@@ -1,0 +1,17 @@
+// Copyright (c) 2020-2022 Navibyte (https://navibyte.com). All rights reserved.
+// Use of this source code is governed by a “BSD-3-Clause”-style license that is
+// specified in the LICENSE file.
+//
+// Docs: https://github.com/navibyte/geospatial
+
+/// Geographic and projected positions and bounding boxes.
+///
+/// This libary exports a subset of `package:geobase/geobase.dart`.
+///
+/// Usage: import `package:geobase/coordinates.dart`
+library coordinates;
+
+export 'src/coordinates/base.dart';
+export 'src/coordinates/geographic.dart';
+export 'src/coordinates/projected.dart';
+export 'src/coordinates/projection.dart';