+++ dart/geobase/lib/meta.dart
@@ -1,0 +1,15 @@
+// Copyright (c) 2020-2022 Navibyte (https://navibyte.com). All rights reserved.
+// Use of this source code is governed by a “BSD-3-Clause”-style license that is
+// specified in the LICENSE file.
+//
+// Docs: https://github.com/navibyte/geospatial
+
+/// Temporal data structures (instant, interval) and spatial extents.
+///
+/// This libary exports a subset of `package:geobase/geobase.dart`.
+///
+/// Usage: import `package:geobase/meta.dart`
+library meta;
+
+export 'src/meta/extent.dart';
+export 'src/meta/time.dart';