+++ dart/geobase/lib/projections_proj4d.dart
@@ -6,6 +6,21 @@
 
 /// Projections provided by the external `proj4dart` package.
 ///
+/// This package implements a projection adapter that acts as a wrapper and uses
+/// internally the projection and coordinate reference system support of the
+/// `proj4dart` package to which there's a dependency.
+///
+/// Usage: import `package:geobase/projections_proj4d.dart`
+///
+/// You might want to import also `package:geobase/geobase.dart` providing base
+/// classes.
+library projections_proj4d;
+
+export 'src/projections_ext/proj4d.dart';
+
+/*
+/// Projections provided by the external `proj4dart` package.
+///
 /// The `geobase` package implements a projection adapter that acts as a wrapper
 /// and uses internally the projection and coordinate reference system support
 /// of the `proj4dart` package to which there's a dependency.
@@ -21,3 +36,4 @@ export 'package:geobase/geobase.dart';
 // export also `proj4dart` specific classes.
 // ignore: directives_ordering
 export 'src/projections_ext/proj4d.dart';
+*/