+++ dart/geobase/test/vector/writer_test.dart
@@ -6,7 +6,9 @@
 
 // ignore_for_file: lines_longer_than_80_chars
 
-import 'package:geobase/geobase.dart';
+import 'package:geobase/codes.dart';
+import 'package:geobase/coordinates.dart';
+import 'package:geobase/vector.dart';
 
 import 'package:test/test.dart';